Refactor code to handle non integer position

+++ src/main/scala/entity/creature/Alliance.scala
@@ -1,6 +1,6 @@
 package entity.creature
 
 abstract class Alliance extends Creature:
-  val maxHealth: Int
-  var health: Int
+  val maxHealth: Double
+  var health: Double
